chore: remove redundant old implementation
